fix Coundown Calendar
Add a call to root.mainloop() and add a file events.txt.  It was able to run in IDLE and mu-editor without the call to mainloop() but it's not clear why it would be able without that call.  I see that the book did not have that call which is a mistake of the book.