Fix the caterpillar game
The caterpillar game was going immediately to game_over because of the missing parentheses on "if outside_window".  Needs to be "if outside_window()".
I also found that the change of direction in response to the arrow keys was not working as it should.
Added a,w,s,d key handling.
Changed how much it increases speed with each leaf